fixing float on ios and eos

+++ napalm/base/test/models.py
@@ -31,7 +31,7 @@ interface = TypedDict(
         "description": str,
         "last_flapped": float,
         "mtu": int,
-        "speed": int,
+        "speed": float,
         "mac_address": str,
     },
 )
